+++ src/Resources/Moderations.php
@@ -15,7 +15,7 @@ final class Moderations implements ModerationsContract
     /**
      * Classifies if text violates OpenAI's Content Policy.
      *
-     * @see https://beta.openai.com/docs/api-reference/moderations/create
+     * @see https://platorm.openai.com/docs/api-reference/moderations/create
      *
      * @param  array<string, mixed>  $parameters
      */