+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v0.6.3 (2023-07-07)
+### Fixed
+- Breaking change on OpenAI API regarding "transient" field in Audio ([#160](https://github.com/openai-php/client/pull/160))
+
+## v0.6.2 (2023-06-23)
+### Changed
+- Error handling: use error code as exception message if error message is empty ([#150](https://github.com/openai-php/client/pull/150))
+
+### Fixed
+- Error handling: Catch error in stream responses ([#150](https://github.com/openai-php/client/pull/150))
+- Error handling: Handle errors where message is an array ([#150](https://github.com/openai-php/client/pull/150))
+
 ## v0.6.1 (2023-06-15)
 ### Fixed
 - Chat/CreateResponse faking with function_call ([#145](https://github.com/openai-php/client/issues/145))
@@ -92,7 +104,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## v0.2.0 (2022-11-07)
 ### Added
-- Add `images()` resource to interact with [DALL-E](https://beta.openai.com/docs/api-reference/images)
+- Add `images()` resource to interact with [DALL-E](https://platform.openai.com/docs/api-reference/images)
 
 ### Fixed
 - Parse completions create response with logprobs correctly